sml_list: make DZG meter workaround more specific
After discussions in issue #105 I asked DZG, and they confirmed
the bug was fixed in firmware version 1.2 and 1.3, which ship
with serial numbers starting with "1 DZG 00 6.." (in decimal).

Compare the serial number, and only apply the workaround if it
is indeed less than

  static const unsigned char dzg_serial_fixed[] =
    {0x0a, 0x01, 'D', 'Z', 'G', 0x00, 0x03, 0x93, 0x87, 0x00};

which is "1 DZG 00 60000000" in their decimal representation.

This fixes #105.