Make sml_workarounds a bitmap, expose in public API
The struct `sml_workarounds` has become a bitmap now, and it is part of
the public API.

I rephrased the DZG workaround explanation a bit to match the new
information: The workaround can't be applied automatically because we
need meter model *and* serial number, but only get the latter via SML.
The serial numbers that have the fix apparently differ between DZG meter
models.

This is also why `sml_workarounds` needs to be public: There needs to
be a way for the libsml user to configure at least this workaround for
the meter at hand. Implementing it as a bitmap is memory-efficient in
contrast to passing an additional struct around everywhere.

Adding `sml_workarounds` to `sml_buffer` makes it possible to configure
workarounds individually for multiple meters that may be handled by the
user, each of which will have their own buffer.